requirements.txt
Django==4.1.3
djangorestframework==3.14.0
future==0.18.2
gunicorn==20.1.0
mime==0.1.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
